+++ R/inferences_conformal.R
@@ -123,7 +123,17 @@ conformal_cv_plus <- function(x, test, R, score, conf_level, mfx = NULL, ...) {
     for (i in idx) {
         data_cv <- train[-i, ]
         # re-fit the original model on training sets withholding the CV fold
-        model_cv <- stats::update(mfx@model, data = data_cv)
+        model_cv <- tryCatch(stats::update(mfx@model, data = data_cv),
+            error = function(e) NULL)
+        if (is.null(model_cv)) {
+            if (is.call(mfx@call_model) && "data" %in% names(mfx@call_model)) {
+                # if the model call has a data argument, we can update it
+                mfx@call_model$data <- data_cv
+                model_cv <- eval(mfx@call_model)
+            } else {
+                stop_sprintf("Failed to re-fit the model on the cross-validation set.")
+            }
+        }
         # use the updated model to make out-of-fold predictions
         # call_cv is the `predictions()` call, which we re-evaluate in-fold: newdata=train[i,]
         call_cv <- mfx@call